+++ Templates/${fileName}.ts.template.js
@@ -1,1 +1,2 @@
+/* The placeholder __componentName__ will be replaced with the actual component name in the `processAfterEach` hook. */
 module.exports = () => `export * from './__componentName__';`;